fix: ``.getTransactions()`` support for complete and correct sorted access
Transactions are correctly be sorted between backends, accounts.
Internal requests are parallel when possible or useful.  Usage is an
AsyncGenerator that allows client app to request as many as needed to
implement any type of paging or infinite scrolling independantly from
the implementation of the backends that can focus on the best strategy
to optimize requesting.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>